Platform/DataHubCpu: Do not delete the `nvda_drv` NVRAM variable
a1914f1 added support for setting the `nvda_drv` NVRAM variable due to
changes in the Sierra Web Drivers (`nvda_drv` is no longer a boot arg).
However, it also deleted the variable if the NvidiaWeb option was not
set in config.plist, removing the choice selected in the NVIDIA Driver
Manager. Do not delete the `nvda_drv` NVRAM variable if NvidiaWeb is
not set in config.plist.